+++ calplot/__init__.py
@@ -1,5 +1,5 @@
-__version_info__ = ('0', '1', '7', '2')
-__date__ = '7 Mar 2021'
+__version_info__ = ('0', '1', '7', '3')
+__date__ = '15 Mar 2021'
 
 __version__ = '.'.join(__version_info__)
 __author__ = 'Tom Kwok'